requirements.txt
torch>=2.4.1
transformers>=4.30
tensorboardX
pyyaml
tqdm